execution-core: Rename lib to `dusk-core`
